Add confirm delete dialog

+++ src/features/pageItems/ItemContextMenu.tsx
@@ -4,6 +4,8 @@ import ContentCopyIcon from '@mui/icons-material/ContentCopy'
 import EditIcon from '@mui/icons-material/Edit'
 import DeleteIcon from '@mui/icons-material/Delete'
 import { useStore } from 'hooks/useStore'
+import { ConfirmDialog } from './ConfirmDialog'
+import { useState } from 'react'
 
 export type ContextMenu = {
 	left: number
@@ -20,6 +22,7 @@ export const ItemContextMenu = ({
 	setContextPos,
 }: ItemContextMenuProps) => {
 	const { appStore } = useStore()
+	const [openConfirmDialog, setOpenConfirmDialog] = useState(false)
 
 	function onClose() {
 		setContextPos({ left: 0, top: 0 })
@@ -44,45 +47,57 @@ export const ItemContextMenu = ({
 	}
 
 	function onDelete() {
-		appStore.delete(appStore.selectedItem.id)
+		setOpenConfirmDialog(true)
 		onClose()
 	}
 
 	return (
-		<Menu
-			open={Boolean(contextPos.left && contextPos.top)}
-			onClose={onClose}
-			anchorReference='anchorPosition'
-			anchorPosition={contextPos}
-			disableAutoFocusItem
-			slotProps={{ paper: { sx: { boxShadow: 2 } } }}
-		>
-			{appStore.selectedList === 'mangas' && (
-				<MenuItem onClick={onGoogleSearch}>
+		<>
+			<Menu
+				open={Boolean(contextPos.left && contextPos.top)}
+				onClose={onClose}
+				anchorReference='anchorPosition'
+				anchorPosition={contextPos}
+				disableAutoFocusItem
+				slotProps={{ paper: { sx: { boxShadow: 2 } } }}
+			>
+				{appStore.selectedList === 'mangas' && (
+					<MenuItem onClick={onGoogleSearch}>
+						<ListItemIcon>
+							<SearchIcon fontSize='small' />
+						</ListItemIcon>
+						<ListItemText>Google Search</ListItemText>
+					</MenuItem>
+				)}
+				<MenuItem onClick={onCopy}>
 					<ListItemIcon>
-						<SearchIcon fontSize='small' />
+						<ContentCopyIcon fontSize='small' />
 					</ListItemIcon>
-					<ListItemText>Google Search</ListItemText>
+					<ListItemText>Copy</ListItemText>
 				</MenuItem>
-			)}
-			<MenuItem onClick={onCopy}>
-				<ListItemIcon>
-					<ContentCopyIcon fontSize='small' />
-				</ListItemIcon>
-				<ListItemText>Copy</ListItemText>
-			</MenuItem>
-			<MenuItem onClick={onEdit}>
-				<ListItemIcon>
-					<EditIcon fontSize='small' />
-				</ListItemIcon>
-				<ListItemText>Edit</ListItemText>
-			</MenuItem>
-			<MenuItem onClick={onDelete}>
-				<ListItemIcon>
-					<DeleteIcon fontSize='small' />
-				</ListItemIcon>
-				<ListItemText>Delete</ListItemText>
-			</MenuItem>
-		</Menu>
+				<MenuItem onClick={onEdit}>
+					<ListItemIcon>
+						<EditIcon fontSize='small' />
+					</ListItemIcon>
+					<ListItemText>Edit</ListItemText>
+				</MenuItem>
+				<MenuItem onClick={onDelete}>
+					<ListItemIcon>
+						<DeleteIcon fontSize='small' />
+					</ListItemIcon>
+					<ListItemText>Delete</ListItemText>
+				</MenuItem>
+			</Menu>
+			<ConfirmDialog
+				open={openConfirmDialog}
+				onConfirm={() => {
+					appStore.delete(appStore.selectedItem.id)
+					setOpenConfirmDialog(false)
+				}}
+				onCancel={() => setOpenConfirmDialog(false)}
+				title='Delete'
+				content={`Are you sure you want to delete ${appStore.selectedItem.name}?`}
+			/>
+		</>
 	)
 }
